+++ src/mplchart/primitives/axes.py
@@ -29,7 +29,7 @@ class NewAxes(Primitive):
             return NotImplemented
 
         result = copy.copy(indicator)
-        result.default_pane = self.target
+        result.target_pane = self.target
         return result
 
     def plot_handler(self, prices, chart, ax=None):