+++ src/mplchart/primitives/price.py
@@ -1,4 +1,6 @@
-""" Price primitive """
+"""Price primitive"""
+
+from matplotlib import pyplot as plt
 
 from ..model import Primitive
 from ..utils import series_data, series_xy
@@ -18,8 +20,17 @@ class Price(Primitive):
         Price('open')   # plot the open price series
     """
 
-    def __init__(self, item: str = 'close', *, color: str = None):
+    def __init__(
+        self,
+        item: str = "close",
+        *,
+        width: float = 1.0,
+        alpha: float = 1.0,
+        color: str = None,
+    ):
         self.item = item
+        self.width = width
+        self.alpha = alpha
         self.color = color
 
     def plot_handler(self, prices, chart, ax=None):
@@ -29,8 +40,12 @@ class Price(Primitive):
         data = series_data(prices, self.item)
         data = chart.extract_df(data)
 
+        textcolor = plt.rcParams["text.color"]
+
         label = self.item
-        color = self.color
+        width = self.width
+        alpha = self.alpha
+        color = self.color or textcolor
 
         xv, yv = series_xy(data)
-        ax.plot(xv, yv, label=label, color=color)
+        ax.plot(xv, yv, label=label, linewidth=width, alpha=alpha, color=color)